Move cmd/dockerd/trap to daemon/command/trap
