+++ daemon/command/trap/trap.go
