+++ daemon/command/trap/trap_linux_test.go
